Add command line switch to enable pdb trace / stack trace on signal
When the --dbg_on_sigusr argument is passed, we install two signal
handlers for debugging.

SIGUSR1 - Enters the Python debugger. Currently it keeps printing the
message "QCoreApplication::exec: The event loop is already running"
because it breaks in the event loop but events keep coming in. The
debugger works, but the constant output is annoying. I'll check if there
is a workaround for this.

SIGUSR2 - Simply prints a stack trace.